Nuevo ComentarioFiltroDto, UsuarioFlitroDto: Modificaciones en ComentariosController,NotificacionDto,ComentarioService,IComentarioService, IUsuarioService

+++ BlogApi/Mapper/Mapper.cs
@@ -95,7 +95,7 @@ public static class Mapper
         {
             Id = n.Id,
             UsuarioDestinoId = n.UsuarioDestinoId,
-            UsuarioOrigenId = (int)n.UsuarioOrigenId,
+            UsuarioOrigenId = n.UsuarioOrigenId,
             Tipo = n.Tipo,
             Mensaje = n.Mensaje,
             FechaCreacion = n.FechaCreacion,
